+++ tests/testthat/test-exclude_nca.R
@@ -23,7 +23,7 @@ test_that("exclude_nca", {
   expect_equal(as.data.frame(my_result_excluded)$exclude,
                c(rep(NA_character_, 4),
                  rep("Span ratio < 1", 10)))
-  
+
   my_result_excluded <- exclude(my_result, FUN=exclude_nca_min.hl.r.squared())
   expect_equal(as.data.frame(my_result_excluded)$exclude,
                c(rep(NA_character_, 4),
@@ -32,7 +32,7 @@ test_that("exclude_nca", {
   expect_equal(as.data.frame(my_result_excluded)$exclude,
                c(rep(NA_character_, 4),
                  rep("Half-life r-squared < 0.95", 10)))
-  
+
   my_data <- PKNCAdata(my_conc, intervals=data.frame(start=0, end=Inf, cmax=TRUE))
   suppressMessages(
     my_result <- pk.nca(my_data)
@@ -46,4 +46,26 @@ test_that("exclude_nca", {
   expect_equal(my_result,
                exclude(my_result, FUN=exclude_nca_min.hl.r.squared()),
                info="Result is ignored when not calculated")
+})
+
+test_that("exclude_nca_conc_count_measured", {
+  my_conc <- PKNCAconc(data.frame(conc=c(1.1^(c(3:0, -Inf)), 1.1), time=0:5, subject = 1), conc~time|subject)
+  my_data <- PKNCAdata(my_conc, intervals=data.frame(start=0, end=Inf, aucinf.obs=TRUE, aucpext.obs=TRUE, count_conc_measured = TRUE))
+  suppressMessages(
+    my_result <- pk.nca(my_data)
+  )
+  expect_equal(
+    as.data.frame(my_result)$exclude,
+    rep(NA_character_, 15)
+  )
+  my_result_exclude5 <- exclude(my_result, FUN = exclude_nca_conc_count_measured(min_count = 5))
+  expect_equal(
+    as.data.frame(my_result_exclude5)$exclude,
+    rep(NA_character_, 15)
+  )
+  my_result_exclude10 <- exclude(my_result, FUN = exclude_nca_conc_count_measured(min_count = 10))
+  expect_equal(
+    as.data.frame(my_result_exclude10)$exclude,
+    c("Number of measured concentrations is < 10", rep(NA_character_, 12), rep("Number of measured concentrations is < 10", 2))
+  )
 })